requirements.txt
certifi
chardet
idna
python-box
PyYAML
requests
tabulate
urllib3
